+++ Libplanet/Blockchain/BlockChain.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Threading;
 using Bencodex.Types;
 using Libplanet.Action;
+using Libplanet.Action.Loader;
 using Libplanet.Assets;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Blockchain.Renderers;
@@ -139,7 +140,7 @@ namespace Libplanet.Blockchain
                     _ => policy.BlockAction,
                     blockChainStates: blockChainStates,
                     genesisHash: genesisBlock.Hash,
-                    actionTypeLoader: StaticActionLoader.Create<T>(),
+                    actionTypeLoader: new SingleActionLoader(typeof(T)),
                     feeCalculator: null
                 )
             )